Opened in VS Code
Welp it seems like things have changed and I won't be able to do git from the full VS Studio. Or I can? But it'll take me a while to remember how to do it and this is MUCH simpler. So basically follow the buttons and it tells you how to set everything up yeet